More robust speediness (#26)
* More robust speediness

* Finish comment

* Remove opt in

I think this can actually always with avoided as long as people are precise with the type names in their asset bundle

* Move around readme

Making this change because the plugin readme should probably go first.

If there is ever a proper modding docs setup I will write more detailed notes on the individual patches.

* Reword comment

* Update README

Clarified the description of method replacements for better understanding.